Moved helpers into appropriate modules and updated notebooks.
