+++ src/segtraq/SegTraQ.py
@@ -1460,7 +1460,7 @@ class _RSFacade:
         min_genes: int = 5,
         pseudocount: float = 0.5,
         n_permutations: int = 200,
-        random_state: int | None = None,
+        random_state: int | None = 42,
         n_jobs: int | None = None,
         parallel_backend: str = "threading",
         inplace: bool = True,
@@ -1750,6 +1750,10 @@ class _CSFacade:
         random_state: int = 42,
         cell_type_key: str | None = None,
         use_hvg: bool | None = None,
+        exclude_gene_prefixes: tuple[str, ...] = ("MT-", "RPL", "RPS"),
+        n_neighbors: int = 15,
+        n_pcs: int = 50,
+        target_sum: float | None = None,
         inplace: bool = True,
         leiden_kwargs: dict | None = None,
     ) -> float:
@@ -1762,6 +1766,10 @@ class _CSFacade:
             random_state=random_state,
             cell_type_key=cell_type_key,
             use_hvg=use_hvg,
+            exclude_gene_prefixes=exclude_gene_prefixes,
+            n_neighbors=n_neighbors,
+            n_pcs=n_pcs,
+            target_sum=target_sum,
             inplace=inplace,
             leiden_kwargs=leiden_kwargs,
         )
@@ -1774,6 +1782,10 @@ class _CSFacade:
         frac_cells_subset: float = 0.63,
         key_prefix: str = "leiden_subset",
         use_hvg: bool | None = None,
+        exclude_gene_prefixes: tuple[str, ...] = ("MT-", "RPL", "RPS"),
+        n_neighbors: int = 15,
+        n_pcs: int = 50,
+        target_sum: float | None = None,
         inplace: bool = True,
         leiden_kwargs: dict | None = None,
     ) -> float:
@@ -1784,6 +1796,10 @@ class _CSFacade:
             tables_key=self._p.tables_key,
             key_prefix=key_prefix,
             use_hvg=use_hvg,
+            exclude_gene_prefixes=exclude_gene_prefixes,
+            n_neighbors=n_neighbors,
+            n_pcs=n_pcs,
+            target_sum=target_sum,
             inplace=inplace,
             leiden_kwargs=leiden_kwargs,
         )
@@ -1796,6 +1812,10 @@ class _CSFacade:
         frac_cells_subset: float = 0.63,
         key_prefix: str = "leiden_subset",
         use_hvg: bool | None = None,
+        exclude_gene_prefixes: tuple[str, ...] = ("MT-", "RPL", "RPS"),
+        n_neighbors: int = 15,
+        n_pcs: int = 50,
+        target_sum: float | None = None,
         inplace: bool = True,
         leiden_kwargs: dict | None = None,
     ) -> float:
@@ -1806,6 +1826,10 @@ class _CSFacade:
             key_prefix=key_prefix,
             tables_key=self._p.tables_key,
             use_hvg=use_hvg,
+            exclude_gene_prefixes=exclude_gene_prefixes,
+            n_neighbors=n_neighbors,
+            n_pcs=n_pcs,
+            target_sum=target_sum,
             inplace=inplace,
             leiden_kwargs=leiden_kwargs,
         )
@@ -1820,6 +1844,10 @@ class _CSFacade:
         random_state: int = 42,
         cell_type_key: str | None = None,
         use_hvg: bool | None = None,
+        exclude_gene_prefixes: tuple[str, ...] = ("MT-", "RPL", "RPS"),
+        n_neighbors: int = 15,
+        n_pcs: int = 50,
+        target_sum: float | None = None,
         inplace: bool = True,
         leiden_kwargs: dict | None = None,
     ):
@@ -1832,6 +1860,10 @@ class _CSFacade:
             random_state=random_state,
             cell_type_key=cell_type_key,
             use_hvg=use_hvg,
+            exclude_gene_prefixes=exclude_gene_prefixes,
+            n_neighbors=n_neighbors,
+            n_pcs=n_pcs,
+            target_sum=target_sum,
             inplace=inplace,
             leiden_kwargs=leiden_kwargs,
         )
